add configuration manager, remove Terminal layout
serialization is at basic level only.
serialization needs to be converted to SettingsStorage.
