+++ Configuration/ConfigIndicator/IndicatorElementCollection.cs
@@ -1,11 +1,11 @@
-﻿namespace StockSharp.Configuration
-{
-	using System.Configuration;
+﻿using System.Configuration;
 
-	/// <summary>
-	/// Represents the custom message adapters section in a configuration file.
+namespace StockSharp.Configuration.ConfigIndicator
+{
+    /// <summary>
+	/// Represents the custom indicators section in a configuration file.
 	/// </summary>
-	public class ConnectionElementCollection : ConfigurationElementCollection
+	public class IndicatorElementCollection : ConfigurationElementCollection
 	{
 		/// <summary>
 		/// When overridden in a derived class, creates a new <see cref="T:System.Configuration.ConfigurationElement"/>.
@@ -15,7 +15,7 @@
 		/// </returns>
 		protected override ConfigurationElement CreateNewElement()
 		{
-			return new ConnectionElement();
+			return new IndicatorElement();
 		}
 
 		/// <summary>
@@ -27,7 +27,7 @@
 		/// <param name="element">The <see cref="T:System.Configuration.ConfigurationElement"/> to return the key for. </param>
 		protected override object GetElementKey(ConfigurationElement element)
 		{
-			var elem = (ConnectionElement)element;
+			var elem = (IndicatorElement)element;
 			return elem.Type;
 		}
 	}
