+++ Configuration/ConfigManager/ManagerBase.cs
@@ -8,6 +8,7 @@ namespace StockSharp.Configuration.ConfigManager
 	/// </summary>
 	public abstract class ManagerBase : Window, IPersistable
 	{
+        //TODO: convert serialization in Layout_Persist to SettingsStorage
 		public virtual void Load(SettingsStorage storage)
 		{
 		}
