+++ Configuration/ConfigurationPublic.csproj
@@ -101,27 +101,26 @@
     <Compile Include="..\StockSharpAssemblyInfo.cs">
       <Link>Properties\StockSharpAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="CandleElement.cs" />
-    <Compile Include="CandleElementCollection.cs" />
+    <Compile Include="ConfigCandle\CandleElement.cs" />
+    <Compile Include="ConfigCandle\CandleElementCollection.cs" />
     <Compile Include="ConfigManager\Layout\LayoutManager_DocumentWindow.cs" />
     <Compile Include="ConfigManager\Layout\LayoutManager_Persist.cs" />
     <Compile Include="ConfigManager\Layout\LayoutManager_ToolWindow.cs" />
     <Compile Include="ConfigManager\Layout\LayoutManager_Docking.cs" />
     <Compile Include="ConfigManager\ManagerBase.cs" />
-    <Compile Include="ConfigurationConstants.cs" />
-    <Compile Include="ConnectionElement.cs" />
-    <Compile Include="ConnectionElementCollection.cs" />
-    <Compile Include="DiagramElement.cs" />
-    <Compile Include="DiagramElementCollection.cs" />
+    <Compile Include="ConfigManager\ConfigurationConstants.cs" />
+    <Compile Include="ConfigConnection\ConnectionElement.cs" />
+    <Compile Include="ConfigConnection\ConnectionElementCollection.cs" />
+    <Compile Include="ConfigDiagram\DiagramElement.cs" />
+    <Compile Include="ConfigDiagram\DiagramElementCollection.cs" />
     <Compile Include="Extensions.cs" />
-    <Compile Include="IndicatorElement.cs" />
-    <Compile Include="IndicatorElementCollection.cs" />
+    <Compile Include="ConfigIndicator\IndicatorElement.cs" />
+    <Compile Include="ConfigIndicator\IndicatorElementCollection.cs" />
     <Compile Include="ConfigManager\ConfigurationManager.cs" />
     <Compile Include="ConfigManager\FolderManager.cs" />
     <Compile Include="ConfigManager\Layout\LayoutManager.cs" />
     <Compile Include="ConfigManager\OutputManager.cs" />
     <Compile Include="ConfigManager\SettingsManager.cs" />
-    <Compile Include="Properties\Annotations.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StockSharpSection.cs" />
   </ItemGroup>
