--- Terminal/Terminal/Layout/LayoutKeys.cs
@@ -1,17 +1,0 @@
-﻿namespace StockSharp.Terminal.Layout
-{
-	public static class LayoutKey
-	{
-		public const string Chart = "Chart";
-		public const string Log = "Log";
-		public const string Indicator = "Indicator";
-		public const string News = "News";
-		public const string Order = "Order";
-		public const string OrderLog = "OrderLog";
-		public const string Portfolio = "Portfolio";
-		public const string Security = "Security";
-		public const string Strategy = "Strategy";
-		public const string Trade = "Trade";
-		public const string Window = "Window";
-	}
-}
